Split native settings components
